% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_eval.R
\name{oos_eval}
\alias{oos_eval}
\title{Evalutat OOS forecast error}
\usage{
oos_eval(
  x,
  seas_fun,
  by = "-1 month",
  end_short = tsbox:::time_shift(ts_summary(x)$end, by)
)
}
\description{
This should work with every \code{seas_} function in the package. Typically, the
series is shortened by one month. The series is then forecasted and can be
compared to the actual values.
}
\examples{
oos_eval(casualities, seas_daily)
oos_eval(transact, seas_daily)
}
