% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_evals.R
\name{oos_evals}
\alias{oos_evals}
\title{Evalutate OOS forecast errors over multiple periods}
\usage{
oos_evals(x, seas_fun, ...)
}
\description{
Evalutate OOS forecast errors over multiple periods
}
\examples{
oos_evals(casualities, seas_daily)
oos_evals(transact, seas_daily)
}
