% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_daily.R
\name{seas_daily}
\alias{seas_daily}
\title{Seasonal adjustment, using LOESS}
\usage{
seas_daily(
  x,
  h = 35,
  holiday_df = NULL,
  span_scale = scale_factor(x),
  span_trend = 0.25,
  span_week = 0.3,
  span_month = 0.7,
  span_within_year = 0.02,
  transform = c("none", "log"),
  force_positive = FALSE
)
}
\arguments{
\item{x}{ts-boxable time series, an object of class ts, xts, zoo, data.frame, data.table, tbl, tbl_ts, tbl_time, tis, irts or timeSeries.}

\item{h}{forecast horizon.}
}
\description{
A simple daily seasonal adjustment procedure, using LOESS. It is optimized
for Speed \emph{and} accuracy, and should work for a wast range of daily series.
}
\details{
Currently defaults to CH holidays, but these should be usable in other places
as well. Maybe optimized in the future.
}
\examples{

x <- transact
seas_daily(x)

}
\author{
Christoph Sax
}
