% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_dsa.R
\name{seas_dsa}
\alias{seas_dsa}
\title{Seasonal adjustment dsa}
\usage{
seas_dsa(x, h = 40)
}
\arguments{
\item{x}{ts-boxable time series, an object of class ts, xts, zoo, data.frame, data.table, tbl, tbl_ts, tbl_time, tis, irts or timeSeries.}

\item{h}{forecast horizon.}
}
\description{
Daily Seasonal Adjustment, using the \emph{dsa} package from Bundesbank.
}
\details{
https://cran.r-project.org/web/packages/dsa/vignettes/dsa-vignette.html
}
\examples{
if (interactive()) {
  sadj_dsa(casualties)
}
}
\author{
XCSAX
}
