% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_dummy.R
\name{seas_dummy}
\alias{seas_dummy}
\title{Seasonal adjustment}
\usage{
seas_dummy(x, h = 30)
}
\description{
Seasonal adjustment
}
\examples{
# oos_eval(casualities, seas_dummy)
# oos_eval(transact, seas_dummy)
# oos_evals(casualities, seas_dummy)
# oos_evals(transact, seas_dummy)
}
