% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_loess.R
\name{seas_loess}
\alias{seas_loess}
\title{Seasonal adjustment, using Loess (daily or monthly)}
\usage{
seas_loess(x, h = 35)
}
\arguments{
\item{x}{ts-boxable time series, an object of class ts, xts, zoo, data.frame, data.table, tbl, tbl_ts, tbl_time, tis, irts or timeSeries.}

\item{h}{forecast horizon}
}
\description{
Seasonal adjustment, using Loess (daily or monthly)
}
\author{
XCSAX
}
